BlueNRG1 stable: (BlueNRG1[smart phone]-BlueNRG1 ok),(BlueNRG1[smart phone]-BlueNRG2 SDK >= 3.1.0 ok)  and (BlueNRG1[smart phone]-BlueNRG2 SDK < 3.1.0 ok)

+++ src/main/java/com/st/BLUENRG/fwUpgrade/feature/NewImageTUContentFeature.java
@@ -61,9 +61,9 @@ public class NewImageTUContentFeature extends DeviceTimestampFeature {
         });
     }
 
-    public static byte getParamBlueNRG2(Sample s){
+    public static int getParamBlueNRG2(Sample s){
         if(hasValidIndex(s,0))
-            return s.data[0].byteValue();
+            return s.data[0].intValue();
         //else
         return 0;
     }
